+++ NEWS.md
@@ -1,6 +1,7 @@
 # BEDMatrix 1.3.0.9000
 
 * Support subsetting by `NA`.
+* Use [subsette package](https://github.com/agrueneberg/subsette) for subsetting
 
 # BEDMatrix 1.3.0